\name{cppUniData}
\alias{cppUniData}
\title{The C++ code, of this package take to run from the data two main information.
the coefficient in the basis expansion of the functional data, and the inner product between theses basis.
cppUnidata made this task in the univariate case.}
\usage{
  cppUniData(fd)
}
\arguments{
  \item{fd}{the functional data}
}
\value{
  fdData a list that containing the transpose of the
  coefficients matrix and the inner product between these
  basis functions.
}
\description{
  The C++ code, of this package take to run from the data
  two main information. the coefficient in the basis
  expansion of the functional data, and the inner product
  between theses basis. cppUnidata made this task in the
  univariate case.
}

