% 
% TODO: Write Documentation 
% 

\name{growthfd}
\docType{data}
\alias{growthfd}
\title{The growth data}
\description{This data set gives the functional data form of the growth  data from the package  fda. }
\usage{data(growthfd)}
\format{A functional data object, build using a 20 spline basis.}
\source{package fda.}
\keyword{datasets}
